#5703 #5711. FTP storage library. Added storage library testing program.
<==== Mercurial data ====>
Branch: feature_5703_storage

+++ storage_plugins/storage_plugin_test_client/src/main.cpp
@@ -3,8 +3,8 @@
 #include <stdexcept>
 #include <cctype>
 #include <cstring>
+#include <plugins/storage/third_party/third_party_storage.h>
 #include "library.h"
-#include <third_party_storage.h>
 
 #define CPREF ">> "
 
@@ -635,7 +635,7 @@ namespace client
                 goto end;
             }
 
-            uint32_t written = fwrite(buf, 1, bCount, t);
+            size_t written = fwrite(buf, 1, bCount, t);
             free(buf);
             if (written < bCount)
             {
